refactor: move eregs-component-lib util methods out to common utils
