+++ solution/ui/regulations/eregs-component-lib/src/components/BlockingModalTrigger.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script>
-import { EventCodes } from "../utils";
+import { EventCodes } from "utilities/utils";
 
 export default {
     name: "BlockingModalTrigger",
