+++ solution/ui/regulations/utilities/utils.js
@@ -19,7 +19,10 @@ import _random from "lodash/random";
 import _set from "lodash/set";
 import _transform from "lodash/transform";
 
-//import numeral from "numeral";
+const EventCodes = {
+    SetSection: "SetSection",
+    OpenBlockingModal: "OpenBlockingModal",
+};
 
 /**
  * Converts the given Map object to an array of values from the map
@@ -72,19 +75,13 @@ function delay(seconds) {
     });
 }
 
-// function niceNumber(value) {
-//     if (_isNil(value)) return "N/A";
-//     if (_isString(value) && _isEmpty(value)) return "N/A";
-//     return numeral(value).format("0,0");
-// }
-
-// function nicePrice(value) {
-//     if (_isNil(value)) return "N/A";
-//     if (_isString(value) && _isEmpty(value)) return "N/A";
-//     return numeral(value).format("0,0.00");
-// }
-
-// YYYY-MM-DD to MMM DD, YYYY
+/**
+ * Converts date from YYYY-MM-DD to MMM DD, YYYY
+ *
+ * @param {string} kebabDate - date in `YYYY-MM-DD` format
+ *
+ * @returns {string} - date in `MMM DD, YYYY` format
+ */
 const niceDate = (kebabDate) => {
     if (_isNil(kebabDate)) return "N/A";
     if (_isString(kebabDate) && _isEmpty(kebabDate)) return "N/A";
@@ -505,18 +502,249 @@ const getCurrentPageResultsRange = ({ count, page = 1, pageSize }) => {
     return [firstInRange, lastInRange];
 };
 
+const formatResourceCategories = (resources) => {
+    const rawCategories = JSON.parse(
+        document.getElementById("categories").textContent
+    );
+
+    resources
+        .filter((resource) => resource.category.type === "category")
+        .forEach((resource) => {
+            const existingCategory = rawCategories.find(
+                (category) => category.name === resource.category.name
+            );
+
+            if (existingCategory) {
+                if (!existingCategory.supplemental_content) {
+                    existingCategory.supplemental_content = [];
+                }
+                existingCategory.supplemental_content.push(resource);
+            } else {
+                const newCategory = JSON.parse(
+                    JSON.stringify(resource.category)
+                );
+                newCategory.supplemental_content = [resource];
+                newCategory.sub_categories = [];
+                rawCategories.push(newCategory);
+            }
+        });
+
+    const rawSubCategories = JSON.parse(
+        document.getElementById("sub_categories").textContent
+    );
+
+    resources
+        .filter((resource) => resource.category.type === "subcategory")
+        .forEach((resource) => {
+            const existingSubCategory = rawSubCategories.find(
+                (category) => category.name === resource.category.name
+            );
+
+            if (existingSubCategory) {
+                if (!existingSubCategory.supplemental_content) {
+                    existingSubCategory.supplemental_content = [];
+                }
+                existingSubCategory.supplemental_content.push(resource);
+            } else {
+                const newSubCategory = JSON.parse(
+                    JSON.stringify(resource.category)
+                );
+                newSubCategory.supplemental_content = [resource];
+                rawSubCategories.push(newSubCategory);
+            }
+        });
+    const categories = rawCategories.map((c) => {
+        const category = JSON.parse(JSON.stringify(c));
+        category.sub_categories = rawSubCategories.filter(
+            (subcategory) => subcategory.parent.id === category.id
+        );
+        return category;
+    });
+    categories.sort((a, b) => a.order - b.order);
+    categories.forEach((category) => {
+        category.sub_categories.sort((a, b) => a.order - b.order);
+    });
+    return categories;
+};
+
+function flattenSubpart(subpart) {
+    const result = JSON.parse(JSON.stringify(subpart));
+    const subjectGroupSections = subpart.children
+        .filter((child) => child.type === "subject_group")
+        .flatMap((subjgrp) => subjgrp.children)
+        .filter((child) => child.type === "section");
+
+    result.children = result.children
+        .concat(subjectGroupSections)
+        .filter((child) => child.type === "section");
+
+    return result;
+}
+
+const formatDate = (value) => {
+    const date = new Date(value);
+    const options = {
+        year: "numeric",
+        month: "long",
+        day: "numeric",
+        timeZone: "UTC",
+    };
+    const format = new Intl.DateTimeFormat("en-US", options);
+
+    const formattedDate = format.format(date);
+    const splitDate = formattedDate.split(" ");
+
+    if (splitDate[0] && splitDate[0].length > 4) {
+        const month = splitDate[0];
+        const abbrMonth = month.slice(0, 3);
+        splitDate[0] = abbrMonth;
+        return splitDate.join(" ");
+    }
+
+    return formattedDate;
+};
+
+/**
+ * Recursively search through DOM Element and its children and
+ * surround strings that match `highlightString` with <mark> tags
+ *
+ * @param {HTMLElement} element - element to mutate
+ * @param {string} highlightString - string to match
+ */
+function addMarks(element, highlightString) {
+    function escapeRegex(string) {
+        return string.replace(/[/\-\\^$*+?.()|[\]{}]/g, "\\$&");
+    }
+
+    const regex = new RegExp(escapeRegex(highlightString));
+
+    if (element.nodeType === document.TEXT_NODE) {
+        // note `nodeValue` vs `innerHTML`
+        // nodeValue gives inner text without Vue component markup tags;
+        // innerHTML gives text with Vue Component markup tags;
+        // Currently there is only the tooltip <trigger-btn> tag at beginning
+        const text = element.nodeValue;
+        if (text.toUpperCase().indexOf(highlightString.toUpperCase()) !== -1) {
+            // ignore citation node at bottom of section
+            if (element?.parentNode?.className === "citation-node") {
+                return;
+            }
+
+            if (element?.parentNode?.nodeName === "A") {
+                const closestParagraph = element.parentNode.closest("p");
+                if (closestParagraph.className === "citation-node") {
+                    return;
+                }
+            }
+
+            const innerHtmlOfParentNode = element.parentNode.innerHTML;
+            const indexOfText = innerHtmlOfParentNode.indexOf(text);
+            const textToKeep = innerHtmlOfParentNode.slice(0, indexOfText);
+            const textToAlter = innerHtmlOfParentNode.slice(indexOfText);
+            const newText = textToAlter.replace(
+                regex,
+                "<mark class='highlight'>$&</mark>"
+            );
+            element.parentNode.innerHTML = textToKeep + newText;
+            return;
+        }
+    } else if (element.nodeType === document.ELEMENT_NODE) {
+        for (let i = 0; i < element.childNodes.length; i++) {
+            if (element.childNodes[i].nodeName !== "MARK") {
+                addMarks(element.childNodes[i], highlightString);
+            }
+        }
+    }
+}
+
+/**
+ * Retrieve comma-separated list of strings from query param in URL
+ * and highlight those strings on the page using <mark> tags
+ *
+ * @param {Location} location - Location object with information about current location of document
+ * @param {string} paramKey - name of query parameter containing strings to match and highlight
+ */
+const highlightText = (location, paramKey) => {
+    const textToHighlight = getQueryParam(location, paramKey);
+    if (location.hash && textToHighlight) {
+        const elementId = location.hash.replace(/^#/, "");
+        const targetedSection = document.getElementById(elementId);
+        if (targetedSection) {
+            const textArr = textToHighlight.split(",");
+            textArr.forEach((text) => {
+                if (text.length > 1) {
+                    addMarks(targetedSection, text);
+                }
+            });
+        }
+    }
+};
+
+/**
+ * Scroll to top of HTML element while taking
+ * heights of other elements into consideration
+ *
+ * @param {HTMLElement} element - scroll to this element
+ * @param {number} [offsetPx=0] - pixels to offset scroll from top of screen
+ */
+const scrollToElement = (element, offsetPx = 0) => {
+    const position = element.getBoundingClientRect();
+    window.scrollTo(position.x, element.offsetTop - offsetPx);
+};
+
+/**
+ * Trap focus in an element (higher-order function)
+ * adapted from: https://hidde.blog/using-javascript-to-trap-focus-in-an-element/
+ *
+ * @param {HTMLElement} element - element in which to trap focus
+ *
+ * @returns {() => void} - function to remove event listener when invoked
+ */
+function trapFocus(element) {
+    const focusableEls = element.querySelectorAll(
+        'a[href]:not([disabled]), button:not([disabled]), textarea:not([disabled]), input[type="text"]:not([disabled]), input[type="radio"]:not([disabled]), input[type="checkbox"]:not([disabled]), select:not([disabled]), div:not([disabled]), iframe'
+    );
+
+    const firstFocusableEl = focusableEls[0];
+    const lastFocusableEl = focusableEls[focusableEls.length - 1];
+    const KEYCODE_TAB = 9;
+
+    const callbackFunc = (e) => {
+        const isTabPressed = e.key === "Tab" || e.keyCode === KEYCODE_TAB;
+
+        if (!isTabPressed) {
+            return;
+        }
+
+        if (e.shiftKey) {
+            /* shift + tab */ if (document.activeElement === firstFocusableEl) {
+                lastFocusableEl.focus();
+                e.preventDefault();
+            }
+        } /* tab */ else {
+            if (document.activeElement === lastFocusableEl) {
+                firstFocusableEl.focus();
+                e.preventDefault();
+            }
+        }
+    };
+
+    element.addEventListener("keydown", callbackFunc);
+
+    return function () {
+        element.removeEventListener("keydown", callbackFunc);
+    };
+}
 export {
     mapToArray,
     parseError,
     swallowError,
     delay,
-    //niceNumber,
     getQueryParam,
     removeQueryParams,
     addQueryParams,
     getFragmentParam,
     removeFragmentParams,
-    //nicePrice,
     isFloat,
     removeNulls,
     chopRight,
@@ -540,4 +768,12 @@ export {
     stripQuotes,
     getTagContent,
     getCurrentPageResultsRange,
+    EventCodes,
+    formatResourceCategories,
+    flattenSubpart,
+    formatDate,
+    highlightText,
+    addMarks,
+    scrollToElement,
+    trapFocus,
 };
